pyproject.toml
[project]
name = "tech-challenge-01"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.118.0,<0.119.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "tortoise-orm[asyncpg] (>=0.25.1,<0.26.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "numpy (>=2.3.3,<3.0.0)",
    "joblib (>=1.5.2,<2.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = ["ruff (>=0.13.3,<0.14.0)"]

[tool.ruff]
line-length = 100
extend-select = ["I"]
fix = true
